Read Me
The aim this report is to explain an optimization process for maximizing
the thrust production for a supersonic nozzle. The optimization algorithm is
a combination of CFD analysis, artificial neural network training and a
constrained function minimization. At the beginning we will discuss the
initial test case characteristics and the goal of optimization, then the
optimization algorithm will be discussed. Finally, there would be the
results and discussion about different issue during the whole process.